% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_docum.R
\docType{data}
\name{example.seqfile.Rdata}
\alias{example.seqfile.Rdata}
\title{Example SeqSQC file used in vignette.}
\description{
The SeqSQC object is a list of two objects. The first object
\code{gdsfile} is the filepath of the "example.gds" file which
stores the genotype and meta info of the example data merged with
the benchmark data.  The second object \code{QCresult} contains the
data dimensions (# of samples and variants), sample annotation, and
QC results for sample missing rate, sex check, inbreeding outlier
check, IBD check, and population outlier check.
}
\author{
Qian Liu \email{qliu7@buffalo.edu}
}
\keyword{datasets}
