% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{methods}
\name{gdsfile}
\alias{gdsfile}
\alias{gdsfile}
\alias{gdsfile,SeqSQCclass-method}
\title{Accessors for the 'gdsfile' slot of a seqSQCclass object.
the gdsfile slot holds the filepath of the gds file, which contains the genotypes and meta info for samples and variants.}
\usage{
gdsfile(x)

\S4method{gdsfile}{SeqSQCclass}(x)
}
\arguments{
\item{x}{a SeqSQCclass object}
}
\description{
Accessors for the 'gdsfile' slot of a seqSQCclass object.
the gdsfile slot holds the filepath of the gds file, which contains the genotypes and meta info for samples and variants.
}
