% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissingRate.R
\name{plotMissingRate}
\alias{plotMissingRate}
\title{Plot the sample missing rate.}
\usage{
plotMissingRate(seqfile, interactive = FALSE)
}
\arguments{
\item{seqfile}{SeqSQCclass input file with calculated missing rates.}

\item{interactive}{whether to generate interactive plot. Recommend to use \code{interactive = TRUE} if user perform sample QC using an rmarkdown script and output plot to html format.}
}
\description{
Plot sample missing rates calculated for each population.
}
\examples{
load(system.file("extdata", "example.seqfile.Rdata", package="SeqSQC"))
seqfile@gdsfile <- system.file("extdata", "example.gds", package="SeqSQC")
p <- plotMissingRate(seqfile, interactive=FALSE)
p
}
