% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPop.R
\name{plotPop}
\alias{plotPop}
\title{Plot the PC axes for samples.}
\usage{
plotPop(seqfile, pc1 = "EV1", pc2 = "EV2", interactive = FALSE,
  pairedScatter = FALSE)
}
\arguments{
\item{seqfile}{SeqSQC object with PCA results.}

\item{pc1}{the eigenvector on x axis. The default is "EV1" for eigenvector 1.}

\item{pc2}{the eigenvector on y axis. The default is "EV2" for eigenvector 2.}

\item{interactive}{whether to generate interactive plot. Recommend to use \code{interactive = TRUE} if user perform sample QC using an rmarkdown script and output plot to html format.}

\item{pairedScatter}{whether to plot the paired scatterplot for the first 4 PC axes.}
}
\value{
the ggplot or interactive plot (if output is in html format) for principle components.
}
\description{
Plot principle components for all sample.
}
\examples{
load(system.file("extdata", "example.seqfile.Rdata", package="SeqSQC"))
gfile <- system.file("extdata", "example.gds", package="SeqSQC")
seqfile <- SeqSQC(gdsfile = gfile, QCresult = QCresult(seqfile))
p <- plotPop(seqfile, interactive=FALSE)
p
}
\author{
Qian Liu \email{qliu7@buffalo.edu}
}
