% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{methods}
\name{QCresult}
\alias{QCresult}
\alias{QCresult}
\alias{QCresult,SeqSQCclass-method}
\title{Accessors for the 'QCresult' slot of a SeqSQCclass object.
the QCresult slot holds the sample QC results, including a summary of variants and samples, and QC results from missing rate check, sex check, inbreeding outlier check, IBD check and population outlier check.}
\usage{
QCresult(x)

\S4method{QCresult}{SeqSQCclass}(x)
}
\arguments{
\item{x}{a SeqSQCclass object}
}
\description{
Accessors for the 'QCresult' slot of a SeqSQCclass object.
the QCresult slot holds the sample QC results, including a summary of variants and samples, and QC results from missing rate check, sex check, inbreeding outlier check, IBD check and population outlier check.
}
