% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderReport.R
\name{RenderReport}
\alias{RenderReport}
\title{Render the rmarkdown file for generating a sample QC report.}
\usage{
RenderReport(input, output, interactive = TRUE)
}
\arguments{
\item{input}{SeqSQCclass object with QC results.}

\item{output}{a character string to define the file name for the QC report.}

\item{interactive}{whether to generate interative plots in the report. The default is TRUE.}
}
\description{
Function to render the pre-compiled rmarkdown file to generate the sample QC report.
}
