% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{SeqSQCclass-class}
\alias{SeqSQCclass-class}
\title{A data format to store genotype phenotype and sample QC results from SeqSQC.}
\description{
A SeqSQCclass object is a list of two objects. The first object \code{gdsfile} is the filepath of the GDS (discussed in section below) file which stores the genotype information from the original VCF file. The second object \code{QCresult} is a list of sample information and QC results, which include the dimension (# of samples and variants), sample annotation, and QC results for sample missing rate, sex check, inbreeding outlier check, IBD check, and population outlier check.
}
\section{Slots}{

\describe{
\item{\code{gdsfile}}{A character string for the filepath of the GDS file.}

\item{\code{QCresult}}{A list with sample information and sample QC results.}
}}

