% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{SeqSQCclass}
\alias{SeqSQCclass}
\title{SeqSQCclass object Constructor}
\usage{
SeqSQCclass(gdsfile, QCresult = List())
}
\arguments{
\item{gdsfile}{A character string for the filepath of the GDS file.}

\item{QCresult}{A list with sample information and sample QC results.}
}
\description{
SeqSQCclass object Constructor
}
