% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingRate.R
\name{MissingRate}
\alias{MissingRate}
\title{Sample missing rate check with SeqSQCclass input file.
Function to calculate sample missing rate and to identify sample outlier with high missing rate (> 0.1).}
\usage{
MissingRate(seqfile, remove.samples = NULL)
}
\arguments{
\item{seqfile}{SeqSQCclass input file, which includes the merged gds file for study cohort and benchmark.}

\item{remove.samples}{a vector of sample names for removal from missing rate check. Could be problematic samples identified from other QC steps, or user-defined samples.}
}
\value{
a data frame with sample name, sample missing rate, and an indicator of whether the sample has a missing rate greater than 0.1.
}
\description{
Sample missing rate check with SeqSQCclass input file.
Function to calculate sample missing rate and to identify sample outlier with high missing rate (> 0.1).
}
\details{
The value of the outlier column is set to NA for benchmark samples.
}
\examples{
load(system.file("extdata", "example.seqfile.Rdata", package="SeqSQC"))
seqfile@gdsfile <- system.file("extdata", "example.gds", package="SeqSQC")
seqfile <- MissingRate(seqfile, remove.samples=NULL)
res.mr <- seqfile@QCresult$MissingRate
tail(res.mr)
}
\keyword{MissingRate}
