% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIBD.R
\name{plotIBD}
\alias{plotIBD}
\title{Plot the IBD coefficients for all sample pairs.}
\usage{
plotIBD(seqfile, interactive = FALSE)
}
\arguments{
\item{seqfile}{SeqSQCclass input file with IBD results.}

\item{interactive}{whether to generate interactive plot. Recommend to use \code{interactive = TRUE} if user performs sample QC using an rmarkdown script and output plots to html format.}
}
\description{
Plot IBD coefficients (k0 and k1) calculated from autosomal variants for all samples.
}
\examples{
load(system.file("extdata", "example.seqfile.Rdata", package="SeqSQC"))
seqfile@gdsfile <- system.file("extdata", "example.gds", package="SeqSQC")
p <- plotIBD(seqfile, interactive=FALSE)
p
}
