% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInbreeding.R
\name{plotInbreeding}
\alias{plotInbreeding}
\title{Plot the inbreeding coefficients.}
\usage{
plotInbreeding(seqfile, interactive = FALSE, sdcoef = 5)
}
\arguments{
\item{seqfile}{SeqSQCclass input file with inbreeding result.}

\item{interactive}{whether to generate interactive plot. Recommend to use \code{interactive = TRUE} if user perform sample QC using an rmarkdown script and output plot to html format.}

\item{sdcoef}{how many standard deviation we need for identification of inbreeding outliers. The default is 5.}
}
\value{
@return the ggplot or interactive plot (if output is in html format) for inbreeding coefficients.
}
\description{
Plot inbreeding coefficients calculated from autosomal variants for each population.
}
\examples{
load(system.file("extdata", "example.seqfile.Rdata", package="SeqSQC"))
gdsfile(seqfile) <- system.file("extdata", "example.gds", package="SeqSQC")
p <- plotInbreeding(seqfile, interactive=FALSE)
p
}
\author{
Qian Liu \email{qliu7@buffalo.edu}
}
