% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSexCheck.R
\name{plotSexCheck}
\alias{plotSexCheck}
\title{Plot the X chromosome inbreeding coefficients.}
\usage{
plotSexCheck(seqfile, interactive = FALSE)
}
\arguments{
\item{seqfile}{SeqSQCclass input file with sexcheck results.}

\item{interactive}{whether to generate interactive plot. Recommend to use \code{interactive = TRUE} if user perform sample QC using an rmarkdown script and output plot to html format.}
}
\value{
the ggplot or interactive plot (if output is in html format) for sex inbreeding coefficients.
}
\description{
Plot the inbreeding coefficients calculated from variants on chromosome X for each populaton.
}
\examples{
load(system.file("extdata", "example.seqfile.Rdata", package="SeqSQC"))
seqfile@gdsfile <- system.file("extdata", "example.gds", package="SeqSQC")
p <- plotSexCheck(seqfile, interactive=FALSE)
p
}
