% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqSQC-package.R
\docType{package}
\name{SeqSQC-package}
\alias{SeqSQC-package}
\title{Sample Quality Check for NGS Data using SeqSQC package}
\description{
SeqSQC
}
\details{
Sample Quality Check for NGS Data.
}
\author{
Qian Liu
}
