% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{SeqSQCclass-class}
\alias{SeqSQCclass-class}
\alias{SeqSQCclass-class}
\alias{SeqSQCclass}
\alias{SeqSQCclass-class}
\alias{SeqSQCclass-class}
\alias{gdsfile}
\alias{gdsfile,SeqSQCclass-method}
\alias{QCresult}
\alias{QCresult,SeqSQCclass-method}
\title{A data format to store genotype phenotype and sample QC results from SeqSQC.}
\usage{
SeqSQCclass(gdsfile, QCresult = List())

gdsfile(x)

QCresult(x)

\S4method{gdsfile}{SeqSQCclass}(x)

\S4method{QCresult}{SeqSQCclass}(x)
}
\arguments{
\item{gdsfile}{A character string for the filepath of the GDS file.}

\item{QCresult}{A list with sample information and sample QC results.}

\item{x}{an SeqSQCClass object.}
}
\value{
The filepath to the gds file.
}
\description{
A SeqSQCclass object is a list of two objects. The first object \code{gdsfile} is the filepath of the GDS (discussed in section below) file which stores the genotype information from the original VCF file. The second object \code{QCresult} is a list of sample information and QC results, which include the dimension (# of samples and variants), sample annotation, and QC results for sample missing rate, sex check, inbreeding outlier check, IBD check, and population outlier check.

SeqSQCclass object Constructor

Method gdsfile.

Method QCresult.
}
\section{Slots}{

\describe{
\item{\code{gdsfile}}{A character string for the filepath of the GDS file.}

\item{\code{QCresult}}{A list with sample information and sample QC results.}
}}

\examples{
load(system.file("extdata", "example.seqfile.Rdata", package="SeqSQC"))
gdsfile(seqfile)
QCresult(seqfile)
}
